Unite _button and button
