+++ ev3dev/button.py
@@ -28,14 +28,283 @@ import sys
 if sys.version_info < (3,4):
     raise SystemError('Must be using Python 3.4 or higher')
 
+import array
+import time
+import evdev
 from ev3dev import get_current_platform
 
-# Import the Button class, this is platform specific
+try:
+    # This is a linux-specific module.
+    # It is required by the Button() class, but failure to import it may be
+    # safely ignored if one just needs to run API tests on Windows.
+    import fcntl
+except ImportError:
+    print("WARNING: Failed to import fcntl. Button class will be unuseable!")
+
+
+# Import the button filenames, this is platform specific
+from ev3dev import get_current_platform
 platform = get_current_platform()
 
 # Only EV3 and EVB have buttons
 if platform == 'ev3':
-    from ev3dev._platform.ev3 import Button
-
+    from ev3dev._platform.ev3 import (
+        buttons_filename as platform_buttons_filename,
+        evdev_device_name as platform_evdev_device_name
+    )
 elif platform == 'evb':
-    from ev3dev._platform.evb import Button
+    from ev3dev._platform.evb import (
+        buttons_filename as platform_buttons_filename,
+        evdev_device_name as platform_evdev_device_name
+    )
+else:
+    platform_buttons_filename = None
+    platform_evdev_device_name = None
+
+
+class ButtonBase(object):
+    """
+    Abstract button interface.
+    """
+    _state = set([])
+
+    def __str__(self):
+        return self.__class__.__name__
+
+    @staticmethod
+    def on_change(changed_buttons):
+        """
+        This handler is called by `process()` whenever state of any button has
+        changed since last `process()` call. `changed_buttons` is a list of
+        tuples of changed button names and their states.
+        """
+        pass
+
+    def any(self):
+        """
+        Checks if any button is pressed.
+        """
+        return bool(self.buttons_pressed)
+
+    def check_buttons(self, buttons=[]):
+        """
+        Check if currently pressed buttons exactly match the given list.
+        """
+        return set(self.buttons_pressed) == set(buttons)
+
+    @property
+    def evdev_device(self):
+        """
+        Return our corresponding evdev device object
+        """
+        devices = [evdev.InputDevice(fn) for fn in evdev.list_devices()]
+
+        for device in devices:
+            if device.name == self.evdev_device_name:
+                return device
+
+        raise Exception("%s: could not find evdev device '%s'" % (self, self.evdev_device_name))
+
+    def process(self, new_state=None):
+        """
+        Check for currenly pressed buttons. If the new state differs from the
+        old state, call the appropriate button event handlers.
+        """
+        if new_state is None:
+            new_state = set(self.buttons_pressed)
+        old_state = self._state
+        self._state = new_state
+
+        state_diff = new_state.symmetric_difference(old_state)
+        for button in state_diff:
+            handler = getattr(self, 'on_' + button)
+
+            if handler is not None:
+                handler(button in new_state)
+
+        if self.on_change is not None and state_diff:
+            self.on_change([(button, button in new_state) for button in state_diff])
+
+    def process_forever(self):
+        for event in self.evdev_device.read_loop():
+            if event.type == evdev.ecodes.EV_KEY:
+                self.process()
+
+    @property
+    def buttons_pressed(self):
+        raise NotImplementedError()
+
+    def _wait(self, wait_for_button_press, wait_for_button_release, timeout_ms):
+        tic = time.time()
+
+        # wait_for_button_press/release can be a list of buttons or a string
+        # with the name of a single button.  If it is a string of a single
+        # button convert that to a list.
+        if isinstance(wait_for_button_press, str):
+            wait_for_button_press = [wait_for_button_press, ]
+
+        if isinstance(wait_for_button_release, str):
+            wait_for_button_release = [wait_for_button_release, ]
+
+        for event in self.evdev_device.read_loop():
+            if event.type == evdev.ecodes.EV_KEY:
+                all_pressed = True
+                all_released = True
+                pressed = self.buttons_pressed
+
+                for button in wait_for_button_press:
+                    if button not in pressed:
+                        all_pressed = False
+                        break
+
+                for button in wait_for_button_release:
+                    if button in pressed:
+                        all_released = False
+                        break
+
+                if all_pressed and all_released:
+                    return True
+
+                if timeout_ms is not None and time.time() >= tic + timeout_ms / 1000:
+                    return False
+
+    def wait_for_pressed(self, buttons, timeout_ms=None):
+        return self._wait(buttons, [], timeout_ms)
+
+    def wait_for_released(self, buttons, timeout_ms=None):
+        return self._wait([], buttons, timeout_ms)
+
+    def wait_for_bump(self, buttons, timeout_ms=None):
+        """
+        Wait for the button to be pressed down and then released.
+        Both actions must happen within timeout_ms.
+        """
+        start_time = time.time()
+
+        if self.wait_for_pressed(buttons, timeout_ms):
+            if timeout_ms is not None:
+                timeout_ms -= int((time.time() - start_time) * 1000)
+            return self.wait_for_released(buttons, timeout_ms)
+
+        return False
+
+
+class ButtonEVIO(ButtonBase):
+    """
+    Provides a generic button reading mechanism that works with event interface
+    and may be adapted to platform specific implementations.
+
+    This implementation depends on the availability of the EVIOCGKEY ioctl
+    to be able to read the button state buffer. See Linux kernel source
+    in /include/uapi/linux/input.h for details.
+    """
+
+    KEY_MAX = 0x2FF
+    KEY_BUF_LEN = int((KEY_MAX + 7) / 8)
+    EVIOCGKEY = (2 << (14 + 8 + 8) | KEY_BUF_LEN << (8 + 8) | ord('E') << 8 | 0x18)
+
+    _buttons = {}
+
+    def __init__(self):
+        ButtonBase.__init__(self)
+        self._file_cache = {}
+        self._buffer_cache = {}
+
+        for b in self._buttons:
+            name = self._buttons[b]['name']
+            if name not in self._file_cache:
+                self._file_cache[name] = open(name, 'rb', 0)
+                self._buffer_cache[name] = array.array('B', [0] * self.KEY_BUF_LEN)
+
+    def _button_file(self, name):
+        return self._file_cache[name]
+
+    def _button_buffer(self, name):
+        return self._buffer_cache[name]
+
+    @property
+    def buttons_pressed(self):
+        """
+        Returns list of names of pressed buttons.
+        """
+        for b in self._buffer_cache:
+            fcntl.ioctl(self._button_file(b), self.EVIOCGKEY, self._buffer_cache[b])
+
+        pressed = []
+        for k, v in self._buttons.items():
+            buf = self._buffer_cache[v['name']]
+            bit = v['value']
+
+            if bool(buf[int(bit / 8)] & 1 << bit % 8):
+                pressed.append(k)
+
+        return pressed
+
+
+class Button(ButtonEVIO):
+    """
+    EVB Buttons
+    """
+
+    _buttons = {
+            'up': {'name': platform_buttons_filename, 'value': 103},
+            'down': {'name': platform_buttons_filename, 'value': 108},
+            'left': {'name': platform_buttons_filename, 'value': 105},
+            'right': {'name': platform_buttons_filename, 'value': 106},
+            'enter': {'name': platform_buttons_filename, 'value': 28},
+            'backspace': {'name': platform_buttons_filename, 'value': 14},
+        }
+    evdev_device_name = platform_evdev_device_name
+
+    '''
+    These handlers are called by `process()` whenever state of 'up', 'down',
+    etc buttons have changed since last `process()` call
+    '''
+    on_up = None
+    on_down = None
+    on_left = None
+    on_right = None
+    on_enter = None
+    on_backspace = None
+
+    @property
+    def up(self):
+        """
+        Check if 'up' button is pressed.
+        """
+        return 'up' in self.buttons_pressed
+
+    @property
+    def down(self):
+        """
+        Check if 'down' button is pressed.
+        """
+        return 'down' in self.buttons_pressed
+
+    @property
+    def left(self):
+        """
+        Check if 'left' button is pressed.
+        """
+        return 'left' in self.buttons_pressed
+
+    @property
+    def right(self):
+        """
+        Check if 'right' button is pressed.
+        """
+        return 'right' in self.buttons_pressed
+
+    @property
+    def enter(self):
+        """
+        Check if 'enter' button is pressed.
+        """
+        return 'enter' in self.buttons_pressed
+
+    @property
+    def backspace(self):
+        """
+        Check if 'backspace' button is pressed.
+        """
+        return 'backspace' in self.buttons_pressed
